AdGenGPT-V1-Tools/fast-api-wrapper/requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-generativeai
python-multipart
pillow

#  + annotated-doc==0.0.4                                                                                                                                    
#  + click==8.3.1                                                                                                                                            
#  + fastapi==0.128.0                                                                                                                                        
#  + google-ai-generativelanguage==0.6.15                                                                                                                    
#  + google-api-core==2.29.0                                                                                                                                 
#  + google-api-python-client==2.188.0                                                                                                                       
#  + google-auth-httplib2==0.3.0                                                                                                                             
#  + google-generativeai==0.8.6                                                                                                                              
#  + googleapis-common-protos==1.72.0
#  + grpcio==1.76.0
#  + grpcio-status==1.71.2
#  + httplib2==0.31.2
#  + proto-plus==1.27.0
#  + protobuf==5.29.5
#  + pyparsing==3.3.2
#  + python-multipart==0.0.21
#  + starlette==0.50.0
#  + uritemplate==4.2.0
#  + uvicorn==0.40.0